requirements.txt
alembic==1.0.5
atomicwrites==1.2.1
attrs==18.2.0
Click==7.0
coverage==4.5.2
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
more-itertools==4.3.0
pluggy==0.8.0
py==1.7.0
PyJWT==1.7.1
python-dateutil==2.7.5
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.14
Werkzeug==0.14.1
